refactor: move power menu to separate module
