+++ ags/quicksettings/quicksettings.ts
@@ -4,10 +4,11 @@ import type { Binding } from "types/service";
 import { PopupWindow, showModal } from "window";
 import { Sliders } from "./sliders";
 import { Toggles } from "./toggles";
+import { PowerMenu } from "./power-menu";
 
 const battery = await Service.import("battery");
 const hyprland = await Service.import("hyprland");
-const powerMenuVisible = Variable(false);
+const powerMenuReveal = Variable(false);
 
 /** Constructor for the top buttons in the quicksettings menu. */
 const Button = (props: {
@@ -97,7 +98,7 @@ export const Quicksettings = () => {
     Button({
       icon: "system-shutdown-symbolic",
       async onClick() {
-        powerMenuVisible.value = !powerMenuVisible.value;
+        powerMenuReveal.value = !powerMenuReveal.value;
       },
     }),
   ];
@@ -127,7 +128,7 @@ export const Quicksettings = () => {
           }),
         }),
 
-        PowerMenu(),
+        PowerMenu({ reveal: powerMenuReveal }),
 
         Sliders(),
 
@@ -138,124 +139,10 @@ export const Quicksettings = () => {
       self.hook(
         App,
         () => {
-          powerMenuVisible.value = false;
+          powerMenuReveal.value = false;
         },
         "window-toggled",
       );
     },
-  });
-};
-
-const PowerMenu = () => {
-  return Widget.Revealer({
-    hexpand: false,
-    transition: "slide_down",
-    transitionDuration: 150,
-    revealChild: powerMenuVisible.bind(),
-    child: Widget.Box({
-      className: "power-menu",
-      hexpand: true,
-      vexpand: false,
-      vertical: true,
-      children: [
-        Widget.Box({
-          className: "title",
-          children: [
-            Widget.Icon({
-              className: "icon",
-              icon: "system-shutdown-symbolic",
-              size: 20,
-            }),
-            Widget.Label({
-              label: "Power Off",
-            }),
-          ],
-        }),
-        Widget.Box({
-          className: "entries",
-          vertical: true,
-          children: [
-            Widget.Button({
-              className: "entry",
-              hexpand: true,
-              child: Widget.Box({
-                children: [
-                  Widget.Label({
-                    label: "Restart...",
-                  }),
-                ],
-              }),
-              async onClicked() {
-                App.closeWindow("quicksettings");
-
-                const shutdown = await showModal({
-                  title: "Restart",
-                  description: "Are you sure you want to restart the computer?",
-                  noOption: "Cancel",
-                  yesOption: "Restart",
-                  emphasize: "no",
-                });
-
-                if (shutdown) {
-                  await Utils.execAsync("reboot");
-                }
-              },
-            }),
-            Widget.Button({
-              className: "entry",
-              hexpand: true,
-              child: Widget.Box({
-                children: [
-                  Widget.Label({
-                    label: "Power Off...",
-                  }),
-                ],
-              }),
-              async onClicked() {
-                App.closeWindow("quicksettings");
-
-                const shutdown = await showModal({
-                  title: "Power Off",
-                  description: "Are you sure you want to power off the computer?",
-                  noOption: "Cancel",
-                  yesOption: "Power Off",
-                  emphasize: "no",
-                });
-
-                if (shutdown) {
-                  await Utils.execAsync("shutdown now");
-                }
-              },
-            }),
-            Widget.Button({
-              className: "entry",
-              hexpand: true,
-              child: Widget.Box({
-                children: [
-                  Widget.Label({
-                    label: "Log Out...",
-                  }),
-                ],
-              }),
-              async onClicked() {
-                App.closeWindow("quicksettings");
-
-                const shutdown = await showModal({
-                  title: "Log Out",
-                  description: "Are you sure you want log out of this session?",
-                  noOption: "Cancel",
-                  yesOption: "Log Out",
-                  emphasize: "no",
-                });
-
-                if (shutdown) {
-                  await Utils.execAsync("uwsm stop");
-                }
-              },
-            }),
-          ],
-        }),
-      ],
-    }),
   });
 };
